+++ CHANGELOG.md
@@ -116,6 +116,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   #1617
 - The GDI+ plugin’s output (`-T*:cairo:gdiplus`) now sets DPI in the written
   output. #1617
+- The Quartz plugin’s PNG output (`-Tpng:cairo:quartz`) now sets DPI in the
+  written PNG. #1617
 
 ### Fixed